refactor(iputils): Move some util functions from bridge to own package
These functions are general purpose util functions that are meant to
be used outside of the bridge package. This is particularly useful
since importing the bridge package prevents kraftkit from building
natively on MacOS.
