+++ machine/network/bridge/utils.go
@@ -2,52 +2,15 @@ package bridge
 
 import (
 	"context"
-	"encoding/binary"
 	"fmt"
-	"math/big"
 	"net"
 	"time"
 
 	"github.com/erikh/ping"
 	"github.com/vishvananda/netlink"
 	"kraftkit.sh/internal/set"
+	"kraftkit.sh/machine/network/iputils"
 )
-
-// IpToBigInt converts a 4 bytes IP into a 128 bit integer.
-func IPToBigInt(ip net.IP) *big.Int {
-	x := big.NewInt(0)
-	if ip4 := ip.To4(); ip4 != nil {
-		return x.SetBytes(ip4)
-	}
-	if ip6 := ip.To16(); ip6 != nil {
-		return x.SetBytes(ip6)
-	}
-	return nil
-}
-
-// BigIntToIP converts 128 bit integer into a 4 bytes IP address.
-func BigIntToIP(v *big.Int) net.IP {
-	return net.IP(v.Bytes())
-}
-
-// Increases IP address numeric value by 1.
-func IncreaseIP(ip net.IP) net.IP {
-	rawip := IPToBigInt(ip)
-	rawip.Add(rawip, big.NewInt(1))
-	return BigIntToIP(rawip)
-}
-
-// IsUnicastIP returns true if the provided IP address and network mask is a
-// unicast address.
-func IsUnicastIP(ip net.IP, mask net.IPMask) bool {
-	// broadcast v4 ip
-	if len(ip) == net.IPv4len && binary.BigEndian.Uint32(ip)&^binary.BigEndian.Uint32(mask) == ^binary.BigEndian.Uint32(mask) {
-		return false
-	}
-
-	// global unicast
-	return ip.IsGlobalUnicast()
-}
 
 // BridgeIPs returns all the IPs attached to the provided bridge
 func BridgeIPs(bridge *netlink.Bridge) ([]string, error) {
@@ -88,7 +51,7 @@ func AllocateIP(ctx context.Context, ipnet *net.IPNet, iface *net.Interface, bri
 
 search:
 	for {
-		ip = IncreaseIP(ip)
+		ip = iputils.IncreaseIP(ip)
 
 		select {
 		case <-ctx.Done():
@@ -115,7 +78,7 @@ search:
 			continue
 
 		// Skip the broadcast IP address.
-		case !IsUnicastIP(ip, ipnet.Mask):
+		case !iputils.IsUnicastIP(ip, ipnet.Mask):
 			continue
 
 		// Skip allocated IP addresses.
